requirements.txt
ebooklib
beautifulsoup4
fpdf2
Flask 